added href guard for loading

+++ devices/AlliedVision/DummyAlliedVisionCtrl.py
@@ -30,4 +30,4 @@ class AlliedVisionCtrl(CameraAbstract):
 
     def __call__(self) -> np.ndarray:
         h, w = SPECS_DICT[self.model_name]['h'], SPECS_DICT[self.model_name]['w']
-        return np.random.rand(h, w)
+        return np.random.rand(h, w).astype('uint16')
